proposing some minor fix for object filter rendering logic
scenario: if the filter is currently an object filter, when the user selects another field that is also an object filter, the code breaks, as the current rendering logic is not picking up the new model's fields. Adding a hook to capture field prop updates in object filter prevents cases like this to happen.